requirements.txt
flask
python-dotenv
google-generativeai
Pillow
twilio
gunicorn